Finalization
The Contact Page was now made, finalizing the website. There are three link options on the website, however, only one of them will lead to a different page other than the home page. For the contact page there is no submission location of the info, therefore the info is just being submitted nowhere.